requirements.txt
streamlit==1.29.0
lumaai
requests==2.31.0
pillow==10.1.0 